Match MiniMax H3 output to ComfyUI (#162)
Refs Kosinkadink/comfy-vibe-station#246

Refs Kosinkadink/comfy-vibe-station#289

Exact verified head: aefc894271684d3cf237fe428da1536ed7d44763 (parent 38a9e1368b2eb398eb7cf4a91e431d41a50330a3)

Verification status: PASS - validation inputs pin the landed dinkster-evidence main 54401048b0b5c58ceb7fb5b504eca8e52bec8343 (evidence PR #14 merged). The accepted receipt was produced at Dinkster b8c4ac4f37eb18f59a6b2cc6fb0a455fdc53830d (parent 38a9e1368b2eb398eb7cf4a91e431d41a50330a3) against evidence pre-integration c7aaa2b1d26464c2dd3d3da766268fa7b4fd1a1e and ComfyUI 912fca4f39b875a0360f2c5170568176ea813ded: separate positive and negative one-output conditioning calls sharing the native clip and target latent, cosine 0.9999793333140917, PSNR 52.8116866716011 dB, byte-identical audio.

## Squash-ready description

Match MiniMax H3 output to current ComfyUI by preserving reference precision at text conditioning, stream timestep conversion, and velocity-to-denoised conversion. Accept strict scalar or rowwise generic INT8 scales without weakening ConvRot requirements. Validate separate positive and negative H3 conditioning calls using the landed, scrubbed evidence revision.

## TESTED.md row

| Feature | Hardware | Date | Evidence |
| --- | --- | --- | --- |
| MiniMax H3 output parity and generic INT8 scale geometry | RipperPC (Linux, RTX PRO 6000 Blackwell GPU1, CUDA 13.0) plus CPU product gates | 2026-09-23 | [Dinkster PR 162](https://github.com/Kosinkadink/Dinkster/pull/162), [#246 product PASS](https://github.com/Kosinkadink/comfy-vibe-station/issues/246#issuecomment-5789258028), and [#289 product PASS](https://github.com/Kosinkadink/comfy-vibe-station/issues/289#issuecomment-5789258128) |

+++ tests/test_native_arm.py
@@ -188,10 +188,10 @@ def test_dtype_policy_resolves_every_native_dispatch_family(family_id: str) -> N
     ("family_id", "expected"),
     (
         ("dinkster.flux_dev", ("float16", "float32", "float32")),
-        # Text falls back to float32, never float16, even for families
-        # whose bfloat16 diffusion default degrades to float16: T5-class
-        # RMS variance overflows at float16.
-        ("dinkster.minimax_h3", ("float16", "float32", "float16")),
+        # Generic T5-class bfloat16 defaults fall back to float32 because
+        # their RMS variance overflows at float16. H3 explicitly follows
+        # the reference's float16 conditioner policy instead.
+        ("dinkster.minimax_h3", ("float16", "float16", "float16")),
         ("dinkster.minimax_music3", ("float16", "float32", "float32")),
         ("dinkster.qwen_image", ("float16", "float32", "float16")),
         ("dinkster.sdxl", ("float16", "float32", "float32")),
@@ -214,7 +214,7 @@ def test_dtype_policy_auto_falls_back_when_hardware_lacks_native_bfloat16(
     (
         ("dinkster.flux_dev", ("bfloat16", "bfloat16", "bfloat16")),
         ("dinkster.sd15", ("float16", "float32", "bfloat16")),
-        ("dinkster.minimax_h3", ("bfloat16", "bfloat16", "float16")),
+        ("dinkster.minimax_h3", ("bfloat16", "float16", "float16")),
         ("dinkster.minimax_music3", ("bfloat16", "bfloat16", "float32")),
         ("dinkster.qwen_image", ("bfloat16", "bfloat16", "bfloat16")),
         ("dinkster.triposplat", ("bfloat16", "bfloat16", "float16")),
